removing the dependency between items in a dependency list
This was in error because if baz depended on foo and bar, but you just wanted
to run foo, pub would run bar and baz.

In fixing it, I've removed the order guarantee from dependency lists. So, if
baz depends on foo and bar, in that order, bar may now execute before foo. If
you care to fix that, you can add a dep from foo to bar.

TODO: present a convincing case why we should guarantee argument order. I
suspect there's no need.